Initial import / port of pyfiglet and FIGlet
This is a fairly direct port of the Python `pyfiglet`
(https://github.com/pwaller/pyfiglet), and is not particularly R-ish
at the moment.  right-to-left rendering is not working, nor is
line breaks or justification.

I've included only the base set of fonts, but will organise fetching
additional fonts at runtime.